Move MemoryLabelRepository and MemoryLabelChangeRepository to API module
