+++ tmail-backend/jmap/extensions-api/src/main/scala/com/linagora/tmail/james/jmap/label/MemoryLabelChangeRepository.scala
